refactor and optim
