+++ src/controller.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 use prost::Message;
-use std::{sync::Arc, time::Duration};
+use std::{cmp::Ordering, collections::HashMap, sync::Arc, time::Duration};
 use tokio::{
     sync::{mpsc, RwLock},
     time,
@@ -61,7 +61,8 @@ use crate::{
     },
     pool::Pool,
     protocol::sync_manager::{
-        SyncBlockRequest, SyncBlockRespond, SyncBlocks, SyncManager, SyncTxRequest, SyncTxRespond,
+        sync_block_respond, SyncBlockRequest, SyncBlockRespond, SyncBlocks, SyncManager,
+        SyncTxRequest, SyncTxRespond,
     },
     system_config::SystemConfig,
     util::*,
@@ -1189,14 +1190,14 @@ impl Controller {
     ) -> Result<bool, StatusCodeEnum> {
         let old_status = self.get_global_status().await;
         let own_status = self.get_status().await;
-        if status.height > old_status.1.height && status.height >= own_status.height {
+        let global_height = status.height;
+        if global_height > old_status.1.height && global_height >= own_status.height {
             info!(
                 "update global status: origin: {}, height: {}, hash: 0x{}",
                 node,
-                status.height,
+                global_height,
                 hex::encode(status.hash.clone().unwrap().hash)
             );
-            let global_height = status.height;
             self.update_global_status(node.to_owned(), status).await;
             if global_height > own_status.height {
                 self.try_sync_block().await;
@@ -1214,7 +1215,7 @@ impl Controller {
         }
 
         // request block if own height behind remote's
-        if status.height > own_status.height {
+        if global_height > own_status.height {
             self.try_sync_block().await;
         }
 
@@ -1298,9 +1299,7 @@ impl Controller {
                             .get_sync_block_req(current_height, &global_status)
                             .await
                         {
-                            if !controller_clone.get_sync_state().await {
-                                controller_clone.set_sync_state(true).await;
-                            }
+                            controller_clone.set_sync_state(true).await;
                             current_height = sync_req.end_height;
                             controller_clone
                                 .unicast_sync_block(global_address.0, sync_req.clone())
@@ -1316,7 +1315,6 @@ impl Controller {
                     }
                     ChainStep::OnlineStep => {
                         controller_clone.set_sync_state(false).await;
-                        controller_clone.sync_manager.clear().await;
                         return;
                     }
                     ChainStep::BusyState => return,
@@ -1342,9 +1340,7 @@ impl Controller {
                         .get_sync_block_req(current_height, &global_status)
                         .await
                     {
-                        if !self.get_sync_state().await {
-                            self.set_sync_state(true).await;
-                        }
+                        self.set_sync_state(true).await;
                         current_height = sync_req.end_height;
                         self.unicast_sync_block(global_address.0, sync_req.clone())
                             .await
@@ -1359,7 +1355,6 @@ impl Controller {
                 }
                 ChainStep::OnlineStep => {
                     self.set_sync_state(false).await;
-                    self.sync_manager.clear().await;
                     break;
                 }
                 ChainStep::BusyState => unreachable!(),
@@ -1397,7 +1392,214 @@ impl Controller {
     }
 
     pub async fn set_sync_state(&self, state: bool) {
-        let mut wr = self.is_sync.write().await;
-        *wr = state;
+        if self.get_sync_state().await == state {
+            return;
+        }
+        *self.is_sync.write().await = state;
+        if !state {
+            self.sync_manager.clear().await;
+        }
+    }
+
+    pub async fn handle_event(
+        &self,
+        event_task: EventTask,
+        old_status: &mut HashMap<NodeAddress, ChainStatus>,
+    ) {
+        match event_task {
+            EventTask::SyncBlockReq(req, origin) => {
+                let mut block_vec = Vec::new();
+
+                for h in req.start_height..=req.end_height {
+                    if let Ok(block) = get_full_block(h).await {
+                        block_vec.push(block);
+                    } else {
+                        warn!("handle SyncBlockReq failed: get block({}) failed", h);
+                        break;
+                    }
+                }
+
+                if block_vec.len() as u64 != req.end_height - req.start_height + 1 {
+                    let sync_block_respond = SyncBlockRespond {
+                        respond: Some(sync_block_respond::Respond::MissBlock(
+                            self.local_address.clone(),
+                        )),
+                    };
+                    self.unicast_sync_block_respond(origin, sync_block_respond)
+                        .await;
+                } else {
+                    info!(
+                        "send SyncBlockRespond: to origin: {:x}, height: {} - {}",
+                        origin, req.start_height, req.end_height
+                    );
+                    let sync_block = SyncBlocks {
+                        address: Some(self.local_address.clone()),
+                        sync_blocks: block_vec,
+                    };
+                    let sync_block_respond = SyncBlockRespond {
+                        respond: Some(sync_block_respond::Respond::Ok(sync_block)),
+                    };
+                    self.unicast_sync_block_respond(origin, sync_block_respond)
+                        .await;
+                }
+            }
+            EventTask::SyncBlock => {
+                debug!("receive SyncBlock event");
+                let (global_address, global_status) = self.get_global_status().await;
+                let mut own_status = self.get_status().await;
+                // get chain lock means syncing
+                self.set_sync_state(true).await;
+                {
+                    match {
+                        let chain = self.chain.read().await;
+                        chain.next_step(&global_status).await
+                    } {
+                        ChainStep::SyncStep => {
+                            let mut syncing = false;
+                            {
+                                let mut chain = self.chain.write().await;
+                                while let Some((addr, block)) =
+                                    self.sync_manager.remove_block(own_status.height + 1).await
+                                {
+                                    chain.clear_candidate();
+                                    match chain.process_block(block).await {
+                                        Ok((consensus_config, mut status)) => {
+                                            reconfigure(consensus_config)
+                                                .await
+                                                .is_success()
+                                                .unwrap();
+                                            status.address = Some(self.local_address.clone());
+                                            self.set_status(status.clone()).await;
+                                            own_status = status.clone();
+                                            if status.height
+                                                % self.config.send_chain_status_interval_sync
+                                                == 0
+                                            {
+                                                self.broadcast_chain_status(status).await;
+                                            }
+                                            syncing = true;
+                                        }
+                                        Err(e) => {
+                                            if (e as u64) % 100 == 0 {
+                                                warn!("sync block failed: {}", e.to_string());
+                                                continue;
+                                            }
+                                            warn!("sync block failed: {}. set remote misbehavior. origin: {}", NodeAddress::from(&addr), e.to_string());
+                                            let del_node_addr = NodeAddress::from(&addr);
+                                            let _ = self
+                                                .node_manager
+                                                .set_misbehavior_node(&del_node_addr)
+                                                .await;
+                                            if global_address == del_node_addr {
+                                                let (ex_addr, ex_status) =
+                                                    self.node_manager.pick_node().await;
+                                                self.update_global_status(ex_addr, ex_status).await;
+                                            }
+                                            if let Some(range_heights) = self
+                                                .sync_manager
+                                                .clear_node_block(&addr, &own_status)
+                                                .await
+                                            {
+                                                let (global_address, global_status) =
+                                                    self.get_global_status().await;
+                                                if global_address.0 != 0 {
+                                                    for range_height in range_heights {
+                                                        if let Some(reqs) =
+                                                            self.sync_manager.re_sync_block_req(
+                                                                range_height,
+                                                                &global_status,
+                                                            )
+                                                        {
+                                                            for req in reqs {
+                                                                self.unicast_sync_block(
+                                                                    global_address.0,
+                                                                    req,
+                                                                )
+                                                                .await;
+                                                            }
+                                                        }
+                                                    }
+                                                }
+                                            } else {
+                                                syncing = true;
+                                            }
+                                        }
+                                    }
+                                }
+                            }
+                            if syncing {
+                                self.sync_block().await.unwrap();
+                            }
+                        }
+                        ChainStep::OnlineStep => {
+                            self.set_sync_state(false).await;
+                        }
+                        ChainStep::BusyState => unreachable!(),
+                    }
+                }
+            }
+            EventTask::BroadCastCSI => {
+                info!("receive BroadCastCSI event");
+                let status = self.get_status().await;
+
+                let mut chain_status_bytes = Vec::new();
+                status
+                    .encode(&mut chain_status_bytes)
+                    .map_err(|_| {
+                        warn!("process BroadCastCSI failed: encode ChainStatus failed");
+                        StatusCodeEnum::EncodeError
+                    })
+                    .unwrap();
+                let msg_hash = hash_data(&chain_status_bytes);
+
+                #[cfg(feature = "sm")]
+                let crypto = crypto_sm::crypto::Crypto::new(&self.private_key_path);
+                #[cfg(feature = "eth")]
+                let crypto = crypto_eth::crypto::Crypto::new(&self.private_key_path);
+
+                let signature = crypto.sign_message(&msg_hash).unwrap();
+
+                self.broadcast_chain_status_init(ChainStatusInit {
+                    chain_status: Some(status),
+                    signature,
+                })
+                .await
+                .await
+                .unwrap();
+            }
+            EventTask::RecordAllNode => {
+                let nodes = self.node_manager.nodes.read().await.clone();
+                for (na, current_cs) in nodes.iter() {
+                    if let Some(old_cs) = old_status.get(na) {
+                        match old_cs.height.cmp(&current_cs.height) {
+                            Ordering::Greater => {
+                                error!(
+                                    "node status rollbacked: old height: {}, current height: {}. set it misbehavior. origin: {}",
+                                    old_cs.height,
+                                    current_cs.height,
+                                    &na
+                                );
+                                let _ = self.node_manager.set_misbehavior_node(na).await;
+                            }
+                            Ordering::Equal => {
+                                warn!(
+                                    "node status stale: height: {}. delete it. origin: {}",
+                                    old_cs.height, &na
+                                );
+                                if self.node_manager.in_node(na).await {
+                                    self.node_manager.delete_node(na).await;
+                                }
+                            }
+                            Ordering::Less => {
+                                // update node in old status
+                                old_status.insert(*na, current_cs.clone());
+                            }
+                        }
+                    } else {
+                        old_status.insert(*na, current_cs.clone());
+                    }
+                }
+            }
+        }
     }
 }
